+++ resources/js/components/competences/Competence.vue
@@ -25,10 +25,24 @@ export default {
       CompetenceComponentEmitter.$emit('edit', competence)
     },
     deleteCompetence(competence) {
-      axios.delete(`/competences/${competence.id}`).then(response => {
-        console.log(response.data.message)
-        CompetenceComponentEmitter.$emit('deleted')
-      })
+      this.$swal
+        .fire({
+          title: "Está seguro(a)?",
+          text: "Este cambio no se podrá revertir!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Si, eliminar!",
+          cancelButtonText: "Cancelar"
+        }).then(result => {
+          if (result.value) {
+            axios.delete(`/competences/${competence.id}`).then(response => {
+              this.$swal.fire('Eliminada!', response.data.message, 'success')
+              CompetenceComponentEmitter.$emit('deleted')
+            })
+          }
+        })
     }
   }
 };
