+++ resources/js/components/components/Component.vue
@@ -66,9 +66,22 @@ export default {
       ComponentEmitter.$emit('edit', component)
     },
     deleteComponent(component) {
-      axios.delete(`/components/${component.id}`).then(response => {
-        this.$emit('deleted')
-        console.log(response.data.message)
+      this.$swal.fire({
+        title: 'Está seguro(a)?',
+        text: "Este cambio no se podrá revertir!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, eliminar!',
+        cancelButtonText: 'Cancelar'
+      }).then(result => {
+        if (result.value) {
+          axios.delete(`/components/${component.id}`).then(response => {
+            this.$emit('deleted')
+            this.$swal.fire('Eliminado',response.data.message,'success')
+        })
+        }
       })
     }
   }
