+++ resources/js/components/grades/Grade.vue
@@ -83,10 +83,25 @@ export default {
       })
     },
     deleteGrade() {
-        axios.delete(`/grades/${this.grade.id}`, this.grade).then(response=>{
-          console.log(response.data.message)
-          this.$emit('success')
+      this.$swal
+        .fire({
+          title: "Está seguro(a)?",
+          text: "Este cambio no se podrá revertir!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Si, eliminar!",
+          cancelButtonText: "Cancelar"
+        }).then(result => {
+          if (result.value) {
+            axios.delete(`/grades/${this.grade.id}`, this.grade).then(response=>{
+              this.$swal.fire('Eliminado!', response.data.message, 'success')
+              this.$emit('success')
         })
+          }
+        })
+        
     }
   }
 };
