+++ resources/js/components/ovas/Ova.vue
@@ -106,10 +106,24 @@ export default {
       OvaEmitter.$emit('edit', ova)
     },
     deleteOva(ova) {
-      axios.delete(`/ovas/${ova.id}`).then(response => {
-        console.log(response.data.message)
-        this.$emit('deleted')
+      this.$swal.fire({
+        title: 'Está seguro(a)?',
+        text: "Este cambio no se podrá revertir!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, eliminar!',
+        cancelButtonText: 'Cancelar'
+      }).then(result => {
+        if (result.value) {
+          axios.delete(`/ovas/${ova.id}`).then(response => {
+          this.$swal.fire('Eliminado!', response.data.message, 'success')
+          this.$emit('deleted')
       })
+        }
+      })
+      
     },
     detailOva(ova) {
       OvaEmitter.$emit('detail',ova)
